produce_events/requirements.txt
requests
pandas
kafka-python
psycopg2-binary